test(orchestration): fold manifest schema round-trip tests into orchestrationDomain suite
